A final table on the clean waters score per municipaity is ready. Commit the html, rmd, and the csv file.
However, we may later decide to modify the calcualtions. I did not use chemical status for modified areas and I simply removed the water areas that did not have any status assesse (score NA). Instead, we may later decide to repalce all these NA with a score of 50, for example.